refactor: Chess로 board가 하던일을 위임
