+++ src/main/java/softeer2nd/Main.java
@@ -2,18 +2,18 @@ package softeer2nd;
 
 import java.util.Scanner;
 
-import softeer2nd.chess.domain.board.Board;
-import softeer2nd.chess.domain.game.GameManager;
-import softeer2nd.chess.view.InputView;
+import softeer2nd.chess.domain.Chess;
+import softeer2nd.chess.domain.GameManager;
 import softeer2nd.chess.view.ChessView;
+import softeer2nd.chess.view.InputView;
 
 public class Main {
 
 	public static void main(String[] args) {
 		final InputView inputView = new InputView(new Scanner(System.in));
 		final ChessView chessView = new ChessView();
-		final Board board = new Board();
-		GameManager gameManager = new GameManager(inputView, chessView, board);
+		final Chess chess = new Chess();
+		GameManager gameManager = new GameManager(inputView, chessView, chess);
 
 		gameManager.startGame();
 	}
